create category CRUD

+++ app/Http/Controllers/V1/Product/ProductController.php
@@ -7,7 +7,7 @@ use App\Http\Requests\Product\CreateProductRequest;
 use App\Http\Requests\Product\UpdateProductRequest;
 use App\Models\Product\Product;
 use App\Repositories\Product\ProductRepository;
-use Illuminate\Support\Facades\Auth as FacadesAuth;
+use Illuminate\Support\Facades\Auth;
 
 
 
@@ -25,6 +25,10 @@ class ProductController extends Controller
     }
 
     public function store(CreateProductRequest $request){
+        $user = Auth::user();
+        if(! $user){
+            return 'عدم دسترسی';
+        }
         $error = $this->productRepo->store($request);
         if ($error === null){
             return response()->json(['message' => __('product.create,success')] , 201);
@@ -41,7 +45,7 @@ class ProductController extends Controller
     }
 
     public function update(Product $product , UpdateProductRequest $request){
-        $user = FacadesAuth::user();
+        $user = Auth::user();
         if(! $user){
             return 'عدم دسترسی';
         }
@@ -54,6 +58,10 @@ class ProductController extends Controller
     }
 
     public function destroy(Product $product){
+        $user = Auth::user();
+        if(! $user){
+            return 'عدم دسترسی';
+        }
         $error = $this->productRepo->delete($product);
         if ($error === null){
             return response()->json(['message' => __('product.delete,success')] , 200);
